Bugfix: Dates with month "12" were wrongly parsed
Months from string were directly passed to Date constructor.
Those months are in the range 1-12, but Date expects months in the range 0-11.
For dates with a month of 12, it was parsed as "month 0 of the next year"

--

Cherry picked off master @ f24974cf0a3ab6c8f659266d075ece2922765c69

Closes #9